Fixed Web Parsing Error
The original file was pulling from XML, the XML is still available on the AWS URL, however, it hasn't been updated since 2018, triggering the script to get the data from the TS Website.  Unfortunately, the script would break with an exception due a IntraDay Rate parsing error, as the % symbol needed to be stripped.  This script fixes the parsing issues.